backend/my_flask_app/requirements.txt
Flask==2.2.3
Flask-CORS==3.0.10
Flask-JWT-Extended==4.4.4
bcrypt==4.0.1
pytest==7.2.2
pytest-flask==1.2.0